[Enhancement] Clamp replace-batch-rows in signed space and add UT
Address codex P2 review feedback on #73349:

- `primary_key_compaction_replace_batch_rows` is a mutable signed
  CONF_mInt32. The previous threshold computation cast it to size_t
  before clamping, so a negative misconfig wrapped to ~0ULL, never
  tripped the `>=` check, and let the resolver accumulate an entire
  segment into batch_col. Clamp in int32_t space first.

- Tighten the inline comment: "0 disables batching" was misleading.
  With `std::max(1, …)`, both 0 and negatives collapse to 1 (per-chunk
  fallback), and that is now stated explicitly together with the
  reason the clamp must precede the size_t cast.

- Add `test_replace_batch_rows_correctness` UT: sweeps the knob
  through {-1, 0, 1, 5, 1024} on a fresh tablet per iteration, with
  vector_chunk_size lowered to 4 so 12-row segments produce multiple
  chunks. Exercises the negative-clamp, per-chunk, mid-loop-flush,
  and trailing-only-flush paths and asserts identical post-compaction
  row count, rowset shape, and delvec emptiness.

+++ be/src/storage/primary_key_compaction_conflict_resolver.cpp
@@ -48,10 +48,11 @@ Status PrimaryKeyCompactionConflictResolver::execute() {
                 // overhead: each `params.index->replace()` ends with a memtable flush check + lock
                 // round-trip in LakePersistentIndex::replace(). For a 1 M-row segment with the default
                 // 4 K chunk, that is ~250 such calls per segment. Batching N chunks amortises the
-                // per-call setup, vector allocations, and memtable bookkeeping by ~N×. 0 disables
-                // batching (one replace per chunk).
-                const size_t batch_rows_threshold =
-                        std::max<size_t>(1, static_cast<size_t>(config::primary_key_compaction_replace_batch_rows));
+                // per-call setup, vector allocations, and memtable bookkeeping by ~N×. Values <= 1
+                // (including negatives, which would otherwise wrap when cast to size_t) collapse to
+                // a per-chunk threshold of 1, reverting to pre-patch behaviour.
+                const size_t batch_rows_threshold = static_cast<size_t>(
+                        std::max<int32_t>(1, config::primary_key_compaction_replace_batch_rows));
                 for (size_t segment_id = 0; segment_id < segment_iters.size(); segment_id++) {
                     RETURN_IF_ERROR(breakpoint_check());
                     // only hold pkey, so can use larger chunk size
